feature: add network check

+++ app/src/main/java/com/qxy/siegelions/dao/RankingVersionDao.kt
@@ -16,11 +16,11 @@ interface RankingVersionDao {
     fun updateVersion(rankingEntry: RankingVersion?)
 
     @Query("SELECT version FROM ranking_version WHERE active_time < :time ORDER BY active_time DESC limit 1")
-    fun getVersionIdByDate(time: Date): Long
+    fun getVersionIdByDate(time: Date): Int
 
     @Query("SELECT * FROM ranking_version WHERE active_time = :time")
     fun getVersionByDate(time: Date): RankingVersion
 
     @Query("SELECT * FROM ranking_version WHERE id = :vid")
-    fun getVersionById(vid: Long): RankingVersion?
+    fun getVersionById(vid: Int): RankingVersion?
 }
